refactor(cli): split compose delivery and purge dead code
