--- .playwright-mcp/console-2026-08-20T23-03-59-527Z.log
@@ -1,2 +1,0 @@
-[     242ms] [ERROR] Failed to load resource: the server responded with a status of 401 (Unauthorized) @ http://127.0.0.1:48084/api/v1/auth/whoami:0
-[   15358ms] [ERROR] Failed to load resource: the server responded with a status of 403 (Forbidden) @ http://127.0.0.1:48084/api/v1/instance/retention-health:0
