--- .playwright-mcp/console-2026-08-20T23-04-22-101Z.log
@@ -1,1 +1,0 @@
-[      94ms] [ERROR] Failed to load resource: the server responded with a status of 403 (Forbidden) @ http://127.0.0.1:48084/api/v1/instance/retention-health:0
