+++ internal/cli/compose.go
@@ -1,25 +1,16 @@
 package cli
 
 import (
-	"bytes"
 	"context"
 	"crypto/sha256"
 	"encoding/hex"
-	"encoding/json"
 	"errors"
-	"flag"
 	"fmt"
-	"io"
-	"io/fs"
-	"maps"
 	"net/http"
 	"net/url"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"regexp"
-	"runtime"
-	"slices"
 	"sort"
 	"strings"
 	"time"
@@ -140,355 +131,6 @@ func openComposeStack(st *State, ios IO, flags commonFlags, opts composeStackOpt
 }
 
 // ---------------------------------------------------------------------------
-// hikyo run -- <command>
-// ---------------------------------------------------------------------------
-
-func runRun(ctx context.Context, ios IO, args []string) error {
-	// The child's command and ITS flags sit after the first `--`. Split BEFORE
-	// any flag parsing: flag.Parse would otherwise consume `--` and then eat the
-	// child's own flags (`hikyo run -- mycmd --config-only`).
-	sep := slices.Index(args, "--")
-	if sep < 0 {
-		return failf(ExitUsage, "usage: hikyo run [flags] -- <command> [args...]")
-	}
-	hikyoArgs, childArgs := args[:sep], args[sep+1:]
-	if len(childArgs) == 0 {
-		return failf(ExitUsage, "hikyo run: no command after `--`")
-	}
-
-	var (
-		configOnly       bool
-		useHumanSession  bool
-		allowOverrideRaw string
-		projectDir       string
-	)
-	st, flags, err := parseCommon("run", ios, hikyoArgs, func(fs *flag.FlagSet) {
-		fs.BoolVar(&configOnly, "config-only", false, "request the config-only projection: no secrets, a distinct authorized mode")
-		fs.BoolVar(&useHumanSession, "use-human-session", false, "the locked #18 exception: run under the stored human session, gated by a TTY, an enumerated confirmation, and a live disclosure window")
-		fs.StringVar(&allowOverrideRaw, "allow-override", "", "comma-separated keys whose inherited value the fetched value may replace")
-		fs.StringVar(&projectDir, "project-directory", "", "directory to look up hikyo-compose.yaml from (walks up); optional")
-	})
-	if err != nil {
-		return err
-	}
-	if err := flags.checkNoPositionals("run"); err != nil {
-		return err
-	}
-	allowOverride := splitCSV(allowOverrideRaw)
-
-	// The single narrow #18 exception, restated exactly (api-cli-surface ADR line
-	// 96): `run` — and only `run` — may use the stored human session, and only
-	// when ALL of the flag, a TTY, an enumerated confirmation, and the bound
-	// reauth ceremony hold. `render` and `sync` have no human path, so this branch
-	// lives here rather than in resolveMachineTarget.
-	if useHumanSession {
-		cfg, _, err := findComposeConfig(startDir(ios, projectDir))
-		if err != nil {
-			return err
-		}
-		if flags.Auth == "machine" {
-			return failf(ExitRefused, "hikyo run --use-human-session conflicts with --auth=machine")
-		}
-		flags.Auth = "human"
-		return runHumanSession(ctx, ios, st, flags, cfg, childArgs, configOnly, allowOverride)
-	}
-	stack, err := openComposeStack(st, ios, flags, composeStackOptions{projectDir: projectDir, configOnly: configOnly})
-	if err != nil {
-		return err
-	}
-	snapshotBinding, err := stack.newSnapshotBinding([]string{runGenerationKey})
-	if err != nil {
-		return failf(ExitRefused, "hikyo run: snapshot binding: %v", err)
-	}
-	// Flush-before-fetch (ops-spec § 6 ordering rule): pending offline records
-	// reconcile BEFORE the fetch proceeds; a failure refuses the fetch. A run
-	// without config has no state dir, so the stack method is a no-op.
-	if err := stack.flushOffline(ctx); err != nil {
-		return err
-	}
-
-	// Loader-control acknowledgement (compose ADR § "Loader-control keys"): the
-	// config's run block acknowledges by name. Resolved before the fetch so the
-	// offline path can refuse a loader-control key BEFORE it appends any offline
-	// record (finding 6).
-	var ack []string
-	if stack.cfg != nil {
-		ack = stack.cfg.Run.AcknowledgeLoaderControl
-	}
-
-	resp, ferr := stack.fetchDelivery(ctx, ack, "")
-
-	var (
-		fetched map[string]string
-		live    bool
-	)
-	if ferr != nil {
-		f, herr := serveRunOffline(ios, stack.cfg, stack.stateDir, snapshotBinding, ack, ferr)
-		if herr != nil {
-			return herr
-		}
-		fetched = f
-	} else {
-		if stack.cfg != nil {
-			snapshotBinding, err = bindSnapshotDelivery(snapshotBinding, resp)
-			if err != nil {
-				return failf(ExitRefused, "hikyo run: snapshot binding: %v", err)
-			}
-		}
-		// All-or-nothing (compose ADR § "Authorization"): a secret the principal
-		// cannot reveal makes the whole delivery refuse BEFORE anything else. Not
-		// reached under --config-only, whose projection carries no secrets.
-		if !configOnly {
-			if missing := unrevealedSecrets(resp.Keys); len(missing) > 0 {
-				return failf(ExitRefused, "hikyo run: cannot deliver secret(s) %s — %s",
-					strings.Join(missing, ", "), machineRevealOptIn)
-			}
-		}
-		fetched = deliveredValues(resp.Keys)
-		live = true
-	}
-
-	// Loader-control refusal for the LIVE path (the offline path already checked
-	// pre-append inside serveRunOffline).
-	if refused := compose.RefuseUnacknowledged(mapKeys(fetched), ack); len(refused) > 0 {
-		return failf(ExitRefused, "hikyo run: refusing loader-control key(s) %s; acknowledge each by name in the config's `run.acknowledge_loader_control`",
-			strings.Join(refused, ", "))
-	}
-
-	// Merge: fetched wins; a differing collision is a hard error unless named in
-	// --allow-override (compose ADR § "Merge, collisions"). The base is the
-	// SANITIZED parent environment — HIKYO_TOKEN (the workload credential) never
-	// reaches the child (finding 1).
-	merged, _, err := compose.MergeEnv(sanitizedEnviron(), fetched, allowOverride)
-	if err != nil {
-		return &Error{Code: ExitRefused, Err: err}
-	}
-
-	// ARG_MAX preflight (ops-spec § 6): the execve composite bound, refused loud
-	// pre-exec rather than as E2BIG at the wrong layer.
-	if ok, detail := compose.ExecPreflight(merged, childArgs, compose.DefaultArgMax()); !ok {
-		return failf(ExitRefused, "hikyo run: %s; reduce the delivered set or shorten the command", detail)
-	}
-
-	// Snapshot: after a LIVE delivering fetch and only when a config file exists.
-	// Opt-in governs SERVING, not saving — a silent save failure is the silent
-	// fallback the house forbids, so it is a hard error.
-	if live && stack.cfg != nil {
-		if err := saveRunSnapshot(stack.stateDir, snapshotBinding, resp); err != nil {
-			return failf(ExitInternal, "hikyo run: saving offline snapshot: %v", err)
-		}
-	}
-
-	// Exec. 127 = not found, 126 = found-but-not-executable — the child-side
-	// convention (exit.go), the only exits outside the closed set. On success
-	// there is no hikyo process (unix syscall.Exec): the child's status is the
-	// invocation's.
-	command := childArgs[0]
-	resolvedPath, cerr := resolveChildCommand(command)
-	if cerr != nil {
-		return cerr
-	}
-	if err := ios.exec(resolvedPath, childArgs, merged); err != nil {
-		return failf(ExitCommandNotExecutable, "hikyo run: %s: %v", command, err)
-	}
-	// Unreachable on a real unix exec (the process image is replaced); reached
-	// only through the injected test seam, which returns nil to signal capture.
-	return nil
-}
-
-// runHumanSession implements the locked #18 exception for `hikyo run`. All four
-// conditions must hold before the child is launched:
-//
-//  1. the --use-human-session flag (the caller is here, so it is set);
-//  2. stderr is a TTY — an ADDITIONAL refusal, never the control (ADR line 96):
-//     a human session driving a non-interactive process is refused;
-//  3. an enumerated confirmation — the environment and the exact key names to be
-//     injected, printed to the controlling terminal, answered y/N there;
-//  4. the bound reauth ceremony — a live disclosure window for the environment,
-//     opened inline by the TOTP ceremony where the window allows it. Required
-//     under --config-only too: the four conditions are locked as a set.
-//
-// The offline snapshot machinery is deliberately NOT reached on this path: a
-// human-session snapshot served offline later would bypass the machine-only rule
-// the delivery model rests on, so a human-session run never saves one.
-func runHumanSession(ctx context.Context, ios IO, st *State, flags commonFlags, cfg *compose.Config, childArgs []string, configOnly bool, allowOverride []string) error {
-	// (2) TTY gate, first: a refusal that needs no session and no server. Both
-	// halves are required by name: a controlling terminal (the confirmation and
-	// the code are read there) AND stderr being a TTY (the locked condition -
-	// a human session driving a process whose stderr is captured is refused).
-	terminalSession, terminalErr := ios.terminalSession()
-	if terminalErr != nil {
-		return failf(ExitRefused,
-			"hikyo run --use-human-session requires a controlling terminal for the confirmation and reauth ceremony: %v", terminalErr)
-	}
-	if ios.StderrIsTerminal == nil || !ios.StderrIsTerminal() {
-		return failf(ExitRefused,
-			"hikyo run --use-human-session requires stderr to be a terminal; a captured stderr means a non-interactive process, which the human-session exception refuses")
-	}
-
-	client, artifact, resolved, err := authenticatedTarget(st, ios, flags)
-	if err != nil {
-		return err
-	}
-	session, err := requireHumanSession("hikyo run --use-human-session", artifact)
-	if err != nil {
-		return err
-	}
-	if cfg != nil {
-		for _, d := range []struct {
-			dim Dimension
-			val string
-		}{{DimOrg, cfg.Org}, {DimProject, cfg.Project}, {DimEnv, cfg.Environment}} {
-			if err := foldConfigDim(&resolved, d.dim, d.val, composeConfigName); err != nil {
-				return err
-			}
-		}
-	}
-	project, err := projectBase(resolved)
-	if err != nil {
-		return err
-	}
-	org, env := resolved.Get(DimOrg), resolved.Get(DimEnv)
-	if echo := resolved.Echo(); echo != "" {
-		fmt.Fprintf(ios.Stderr, "target: %s [origin %s, artifact human-session]\n", echo, session.Origin)
-	}
-
-	// (4) Bound reauth ceremony: a live disclosure window over the environment,
-	// opened inline (TOTP) where the environment's window allows it; a
-	// 0-window environment is refused with the browser path named. Required
-	// under --config-only too: the exception's four conditions are locked
-	// without a projection carve-out (api-cli-surface ADR).
-	if err := ensureRevealWindow(ctx, client, st, ios, &session, project, env,
-		// The unit is what the run will inject: the environment's secret keys
-		// for the full projection, its config keys under --config-only - the
-		// exception's "enumerated environment/key-set" named as the act is.
-		disclosure{purpose: "reveal", keys: func(ctx context.Context, env string) ([]string, error) {
-			class := apigen.KeyClassificationSecret
-			if configOnly {
-				class = apigen.KeyClassificationConfig
-			}
-			return keyIDsOf(ctx, client, project, env, class, nil)
-		}},
-		failf(ExitAuth, "a live disclosure window is required: run the reveal ceremony first")); err != nil {
-		return err
-	}
-
-	resp, err := fetchDelivery(ctx, client, org, resolved.Get(DimProject), env, configOnly, nil, "")
-	if err != nil {
-		return err
-	}
-	if !configOnly {
-		if missing := unrevealedSecrets(resp.Keys); len(missing) > 0 {
-			return failf(ExitRefused, "hikyo run: cannot deliver secret(s) %s — %s",
-				strings.Join(missing, ", "), machineRevealOptIn)
-		}
-	}
-	fetched := deliveredValues(resp.Keys)
-
-	// (3) Enumerated confirmation on the controlling terminal.
-	names := mapKeys(fetched)
-	sort.Strings(names)
-	prompt := fmt.Sprintf("About to inject %d value(s) into environment %s and exec %q:\n  %s\nProceed",
-		len(names), env, childArgs[0], strings.Join(names, "\n  "))
-	ok, err := terminalSession.ConfirmEnumerated(prompt)
-	if err != nil {
-		return failf(ExitRefused, "reading the confirmation: %v", err)
-	}
-	if !ok {
-		return failf(ExitRefused, "hikyo run --use-human-session: declined at the confirmation")
-	}
-
-	if refused := compose.RefuseUnacknowledged(mapKeys(fetched), runLoaderControlAck(cfg)); len(refused) > 0 {
-		return failf(ExitRefused, "hikyo run: refusing loader-control key(s) %s; acknowledge each by name in the config's `run.acknowledge_loader_control`",
-			strings.Join(refused, ", "))
-	}
-	merged, _, err := compose.MergeEnv(sanitizedEnviron(), fetched, allowOverride)
-	if err != nil {
-		return &Error{Code: ExitRefused, Err: err}
-	}
-	if ok, detail := compose.ExecPreflight(merged, childArgs, compose.DefaultArgMax()); !ok {
-		return failf(ExitRefused, "hikyo run: %s; reduce the delivered set or shorten the command", detail)
-	}
-	command := childArgs[0]
-	resolvedPath, cerr := resolveChildCommand(command)
-	if cerr != nil {
-		return cerr
-	}
-	if err := ios.exec(resolvedPath, childArgs, merged); err != nil {
-		return failf(ExitCommandNotExecutable, "hikyo run: %s: %v", command, err)
-	}
-	return nil
-}
-
-// runLoaderControlAck is the loader-control acknowledgement in force for a run,
-// from the config's run block (empty when there is no config).
-func runLoaderControlAck(cfg *compose.Config) []string {
-	if cfg == nil {
-		return nil
-	}
-	return cfg.Run.AcknowledgeLoaderControl
-}
-
-// serveRunOffline handles a failed run fetch: if it failed as UNAVAILABLE and
-// the stack opted into offline serve, it opens the snapshot, refuses any
-// unacknowledged loader-control key (finding 6, BEFORE any record is written),
-// records one offline disclosure per key BEFORE returning the values, prints the
-// stale line, and returns them. Any other failure (or no opt-in) is surfaced
-// unchanged. The snapshot is bound to run's identity: TargetNames ["__run__"],
-// so a render snapshot cannot be served here (finding 3).
-func serveRunOffline(ios IO, cfg *compose.Config, stateDir string, binding crypto.SnapshotBinding, ack []string, fetchErr error) (map[string]string, error) {
-	if !isUnavailable(fetchErr) {
-		return nil, fetchErr
-	}
-	if cfg == nil || !cfg.Snapshot.OfflineServe {
-		fmt.Fprintln(ios.Stderr, "hikyo run: offline serve is not enabled for this stack; set snapshot.offline_serve: true in hikyo-compose.yaml to serve stale values during an outage")
-		return nil, fetchErr
-	}
-	payload, binding, err := loadOfflineSnapshot(ios, cfg, binding)
-	if err != nil {
-		return nil, err
-	}
-	// Loader-control BEFORE any disclosure record is written (finding 6): an
-	// unacknowledged loader-control key is refused even when serving stale.
-	if refused := compose.RefuseUnacknowledged(rowNames(payload.Rows), ack); len(refused) > 0 {
-		return nil, failf(ExitRefused, "hikyo run: refusing loader-control key(s) %s; acknowledge each by name in the config's `run.acknowledge_loader_control`",
-			strings.Join(refused, ", "))
-	}
-	stamp := payload.GenerationStamps[runGenerationKey]
-	if err := appendOfflineRecords(ios, stateDir, payload.Rows, binding, stamp); err != nil {
-		return nil, failf(ExitInternal, "hikyo run: recording offline disclosure: %v", err)
-	}
-	aad, err := binding.AAD()
-	if err != nil {
-		return nil, failf(ExitInternal, "hikyo run: reading offline snapshot binding: %v", err)
-	}
-	fmt.Fprintf(ios.Stderr, "serving stale from %s, generation %s\n", aad.IssuedAt, stamp)
-	return rowsToValues(payload.Rows), nil
-}
-
-// saveRunSnapshot seals the delivered env plus one run "generation" stamp. The
-// snapshot's TargetNames is ["__run__"] (run holds no render target), so
-// ContextMatches refuses a render snapshot for run and vice versa (finding 3),
-// and its credential binding is the fingerprint of the presented token (R1-3),
-// authenticated in the header — no separate credential record on disk.
-func saveRunSnapshot(stateDir string, binding crypto.SnapshotBinding, resp apigen.DeliveryResponse) error {
-	if _, err := binding.CanonicalAAD(); err != nil {
-		return err
-	}
-	keys, err := loadLocalKeys(stateDir)
-	if err != nil {
-		return err
-	}
-	rows := deliveredRows(resp.Keys)
-	// run's "generation" stamp is over the canonical row set, keyed to the
-	// run generation key (TargetStamp does not grammar-check the name — only
-	// WriteGeneration does — so the __run__ sentinel is legal here).
-	stamp := compose.TargetStamp(keys, runGenerationKey, canonicalRows(rows))
-	payload := compose.SnapshotPayload{Rows: rows, GenerationStamps: map[string]string{runGenerationKey: stamp}}
-	return saveSnapshot(keys, binding, payload)
-}
-
-// ---------------------------------------------------------------------------
 // hikyo compose render|sync|doctor
 // ---------------------------------------------------------------------------
 
@@ -508,656 +150,6 @@ func runCompose(ctx context.Context, ios IO, args []string) error {
 	default:
 		return failf(ExitUsage, "unknown compose verb %q: use render, sync or doctor", sub)
 	}
-}
-
-func runComposeRender(ctx context.Context, ios IO, args []string) (bool, error) {
-	var (
-		configOnly bool
-		projectDir string
-	)
-	st, flags, err := parseCommon("compose render", ios, args, func(fs *flag.FlagSet) {
-		fs.BoolVar(&configOnly, "config-only", false, "request the config-only projection: no secrets")
-		fs.StringVar(&projectDir, "project-directory", "", "directory to look up hikyo-compose.yaml from (walks up)")
-	})
-	if err != nil {
-		return false, err
-	}
-	if err := flags.checkNoPositionals("compose render"); err != nil {
-		return false, err
-	}
-	moved, _, err := composeRenderCore(ctx, ios, st, flags, projectDir, configOnly)
-	return moved, err
-}
-
-// composeRenderCore is the render pipeline, shared by `compose render` and the
-// render step of `compose sync`. It returns whether any target's stamp moved
-// (so sync knows whether to recreate) and the opened stack.
-func composeRenderCore(ctx context.Context, ios IO, st *State, flags commonFlags, projectDir string, configOnly bool) (bool, *composeStack, error) {
-	stack, err := openComposeStack(st, ios, flags, composeStackOptions{
-		projectDir: projectDir, configOnly: configOnly, requireConfig: true,
-	})
-	if err != nil {
-		return false, nil, err
-	}
-	snapshotBinding, err := stack.newSnapshotBinding(stack.cfg.TargetNames())
-	if err != nil {
-		return false, stack, failf(ExitRefused, "compose render: snapshot binding: %v", err)
-	}
-	if stack.runtimeErr != nil {
-		return false, stack, stack.runtimeErr
-	}
-	// The DEFAULT runtime dir MUST be tmpfs-backed or render refuses (compose ADR
-	// § Where plaintext lives; ops-spec § 6). An EXPLICIT runtime_dir is the
-	// operator's accepted disposition (doctor reports `runtime_not_tmpfs` but the
-	// renderer does not block) — the orchestrator's binding call for finding 2.
-	if !stack.explicitRuntime {
-		if err := os.MkdirAll(stack.runtimeDir, 0o700); err != nil {
-			return false, stack, failf(ExitInternal, "compose render: create runtime dir: %v", err)
-		}
-		if ok, terr := compose.IsTmpfs(stack.runtimeDir); terr != nil {
-			return false, stack, failf(ExitInternal, "compose render: checking runtime dir filesystem: %v", terr)
-		} else if !ok {
-			return false, stack, failf(ExitRefused, "compose render: default runtime dir %s is not backed by tmpfs; rendered plaintext must live only on tmpfs — set an explicit `runtime_dir` on tmpfs in %s", stack.runtimeDir, composeConfigName)
-		}
-	}
-	keys, err := loadLocalKeys(stack.stateDir)
-	if err != nil {
-		return false, stack, err
-	}
-
-	w := compose.NewWriter(stack.stateDir, nil)
-	lock, err := w.BeginRender(stack.cfgDir)
-	if err != nil {
-		return false, stack, failf(ExitRefused, "another hikyo compose process holds the lock for %s", stack.slug)
-	}
-	defer lock.Close()
-
-	// 1. Recover incomplete (torn) generations before anything reads them.
-	if err := lock.Recover(stack.runtimeDir); err != nil {
-		return false, stack, failf(ExitInternal, "compose render: recover: %v", err)
-	}
-	// 2. Flush-before-fetch.
-	if err := stack.flushOffline(ctx); err != nil {
-		return false, stack, err
-	}
-	// 3. Cursor: present it only when the full local eligibility test holds.
-	currentStamps, err := compose.CurrentStamps(stack.cfgDir)
-	if err != nil {
-		return false, stack, failf(ExitRefused, "compose render: %v", err)
-	}
-	present := stack.eligibleCursor(currentStamps)
-
-	// The acknowledgement in force for this render is the UNION of every target's
-	// acknowledge_loader_control (#64 audit field). The server records it and
-	// filters nothing; per-target refusal below stays client-side authoritative.
-	resp, ferr := stack.fetchDelivery(ctx, renderAcknowledged(stack.cfg), present)
-	if ferr != nil {
-		moved, err := stack.renderOffline(ctx, ios, lock, keys, snapshotBinding, ferr)
-		return moved, stack, err
-	}
-	if resp.Current {
-		for _, t := range stack.cfg.TargetNames() {
-			fmt.Fprintf(ios.Stderr, "up to date (generation %s)\n", currentStamps[t])
-		}
-		return false, stack, nil
-	}
-	snapshotBinding, err = bindSnapshotDelivery(snapshotBinding, resp)
-	if err != nil {
-		return false, stack, failf(ExitRefused, "compose render: snapshot binding: %v", err)
-	}
-	moved, err := stack.renderApply(ios, lock, keys, snapshotBinding, resp, currentStamps)
-	return moved, stack, err
-}
-
-// renderApply renders each target from a live full delivery. On ANY
-// refusal it writes no generation and does not advance the cursor.
-func (s *composeStack) renderApply(ios IO, lock *compose.RenderLock, keys *crypto.LocalKeys, binding crypto.SnapshotBinding, resp apigen.DeliveryResponse, currentStamps map[string]string) (bool, error) {
-	if _, err := binding.CanonicalAAD(); err != nil {
-		return false, failf(ExitRefused, "compose render: snapshot binding: %v", err)
-	}
-	plan, err := compose.BuildRenderPlan(liveRenderInput(s.cfg, s.configOnly, resp.Keys))
-	if err != nil {
-		return false, failf(ExitInternal, "compose render: %v", err)
-	}
-	if refusals := renderRefusalMessages(plan.Refusals); len(refusals) > 0 {
-		return false, failf(ExitRefused, "hikyo compose render refused; no generation written, cursor not advanced:\n  %s", strings.Join(refusals, "\n  "))
-	}
-
-	// Publish owns generation materialization, the single stamp commit, and GC.
-	// Snapshot + cursor remain post-publish bookkeeping: they are deliberately
-	// not claimed atomic with the filesystem publication.
-	var allRows []compose.SnapshotRow
-	for _, target := range plan.Targets {
-		allRows = append(allRows, target.SnapshotRows...)
-	}
-	published, err := publishRenderPlan(lock, s.runtimeDir, keys, plan)
-	if err != nil {
-		if pendingErr := persistCommittedPublish(s.stateDir, s.runtimeDir, plan, currentStamps, published); pendingErr != nil {
-			return false, failf(ExitInternal, "compose render: publish: %v", errors.Join(err, fmt.Errorf("persist apply-pending: %w", pendingErr)))
-		}
-		return false, failf(ExitInternal, "compose render: publish: %v", err)
-	}
-	finalStamps := published.Stamps
-	moved, lines := publishOutcome(s.runtimeDir, plan, currentStamps, published)
-	if err := persistCommittedPublish(s.stateDir, s.runtimeDir, plan, currentStamps, published); err != nil {
-		return false, failf(ExitInternal, "compose render: persist apply-pending: %v", err)
-	}
-
-	// Snapshot BEFORE cursor: a snapshot is a harmless cache, but a cursor saved
-	// without a snapshot could read "current" after a reboot with nothing to
-	// serve. If the cursor save fails the snapshot still stands and the next
-	// render does a full fetch.
-	if err := saveSnapshot(keys, binding, compose.SnapshotPayload{Rows: allRows, GenerationStamps: finalStamps}); err != nil {
-		return false, failf(ExitInternal, "compose render: save snapshot: %v", err)
-	}
-	if err := s.saveCursor(resp, finalStamps); err != nil {
-		return false, failf(ExitInternal, "compose render: save cursor: %v", err)
-	}
-
-	for _, l := range lines {
-		fmt.Fprintln(ios.Stderr, l)
-	}
-	return moved, nil
-}
-
-// renderOffline renders each target from the last snapshot when the
-// server is unreachable and the stack opted in. Row→key_id now comes from the
-// sealed payload's rows (finding 3: no cleartext sidecar).
-func (s *composeStack) renderOffline(ctx context.Context, ios IO, lock *compose.RenderLock, keys *crypto.LocalKeys, binding crypto.SnapshotBinding, fetchErr error) (bool, error) {
-	_ = ctx
-	if !isUnavailable(fetchErr) {
-		return false, fetchErr
-	}
-	if !s.cfg.Snapshot.OfflineServe {
-		fmt.Fprintln(ios.Stderr, "hikyo compose render: offline serve is not enabled for this stack; set snapshot.offline_serve: true to render from the last snapshot during an outage")
-		return false, fetchErr
-	}
-	payload, binding, err := loadOfflineSnapshot(ios, s.cfg, binding)
-	if err != nil {
-		return false, err
-	}
-	aad, err := binding.AAD()
-	if err != nil {
-		return false, failf(ExitInternal, "compose render: reading offline snapshot binding: %v", err)
-	}
-	plan, err := compose.BuildRenderPlan(offlineRenderInput(s.cfg, s.configOnly, payload.Rows))
-	if err != nil {
-		return false, failf(ExitInternal, "compose render: %v", err)
-	}
-	if refusals := renderRefusalMessages(plan.Refusals); len(refusals) > 0 {
-		return false, failf(ExitRefused, "hikyo compose render (offline) refused; no generation written:\n  %s", strings.Join(refusals, "\n  "))
-	}
-
-	// One offline record per served key, fsynced BEFORE Publish can materialize
-	// any generation. This audit ordering remains outside filesystem publication.
-	var records []compose.OfflineRecord
-	stamps := make(map[string]string, len(plan.Targets))
-	for _, target := range plan.Targets {
-		// The stamp is bound to the target name and must match the value
-		// WriteGeneration computes after the disclosure record is durable.
-		stamp := compose.TargetStamp(keys, target.Name, target.Content)
-		stamps[target.Name] = stamp
-		for _, row := range target.SnapshotRows {
-			id, err := compose.NewRecordID()
-			if err != nil {
-				return false, failf(ExitInternal, "compose render: record id: %v", err)
-			}
-			records = append(records, compose.OfflineRecord{
-				RecordID: id, KeyID: row.KeyID, KeyName: row.Name, Classification: row.Classification,
-				OccurredAt: ios.now().UTC().Format(time.RFC3339), CredentialID: aad.CredentialID,
-				Generation: stamp, ServedFrom: aad.IssuedAt,
-			})
-		}
-	}
-	if err := compose.Append(s.stateDir, records); err != nil {
-		return false, failf(ExitInternal, "compose render: recording offline disclosure: %v", err)
-	}
-
-	currentStamps, err := compose.CurrentStamps(s.cfgDir)
-	if err != nil {
-		return false, failf(ExitRefused, "compose render: %v", err)
-	}
-	published, err := publishRenderPlan(lock, s.runtimeDir, keys, plan)
-	if err != nil {
-		if pendingErr := persistCommittedPublish(s.stateDir, s.runtimeDir, plan, currentStamps, published); pendingErr != nil {
-			return false, failf(ExitInternal, "compose render: publish: %v", errors.Join(err, fmt.Errorf("persist apply-pending: %w", pendingErr)))
-		}
-		return false, failf(ExitInternal, "compose render: publish: %v", err)
-	}
-	for _, target := range plan.Targets {
-		stamp := published.Stamps[target.Name]
-		if stamp != stamps[target.Name] {
-			return false, failf(ExitInternal, "compose render: target %s stamp changed between offline record and publish", target.Name)
-		}
-		fmt.Fprintf(ios.Stderr, "serving stale from %s, generation %s\n", aad.IssuedAt, stamp)
-	}
-	moved, lines := publishOutcome(s.runtimeDir, plan, currentStamps, published)
-	if err := persistCommittedPublish(s.stateDir, s.runtimeDir, plan, currentStamps, published); err != nil {
-		return false, failf(ExitInternal, "compose render: persist apply-pending: %v", err)
-	}
-	for _, l := range lines {
-		fmt.Fprintln(ios.Stderr, l)
-	}
-	return moved, nil
-}
-
-// publishRenderPlan adapts the pure render plan to the one filesystem-owner
-// operation shared by live and offline flows.
-func publishRenderPlan(lock *compose.RenderLock, runtimeDir string, keys *crypto.LocalKeys, plan compose.RenderPlan) (compose.PublishResult, error) {
-	targets := make(map[string][]byte, len(plan.Targets))
-	for _, target := range plan.Targets {
-		targets[target.Name] = target.Content
-	}
-	return lock.Publish(compose.PublishPlan{RuntimeDir: runtimeDir, Keys: keys, Targets: targets})
-}
-
-func publishOutcome(runtimeDir string, plan compose.RenderPlan, currentStamps map[string]string, published compose.PublishResult) (bool, []string) {
-	var lines []string
-	moved := false
-	for _, target := range plan.Targets {
-		stamp := published.Stamps[target.Name]
-		// A changed stamp or re-materialized tmpfs generation both require sync to
-		// re-apply the target (R1-10).
-		if currentStamps[target.Name] != stamp || published.Materialized[target.Name] {
-			moved = true
-			lines = append(lines, fmt.Sprintf("rendered %s generation %s → %s", target.Name, stamp, filepath.Join(runtimeDir, stamp, target.Name+".env")))
-		} else {
-			lines = append(lines, fmt.Sprintf("unchanged %s generation %s", target.Name, stamp))
-		}
-	}
-	return moved, lines
-}
-
-// persistCommittedPublish closes the crash/error window between a stamp switch
-// and sync's Docker apply. Render writes the marker too: a successful render is
-// filesystem-visible but remains unapplied until a later sync clears it.
-func persistCommittedPublish(stateDir, runtimeDir string, plan compose.RenderPlan, currentStamps map[string]string, published compose.PublishResult) error {
-	if !published.CandidateActive() {
-		return nil
-	}
-	moved, _ := publishOutcome(runtimeDir, plan, currentStamps, published)
-	if !moved {
-		return nil
-	}
-	return writeApplyPending(stateDir, published.Stamps)
-}
-
-// ---------------------------------------------------------------------------
-// hikyo compose sync
-// ---------------------------------------------------------------------------
-
-func runComposeSync(ctx context.Context, ios IO, args []string) error {
-	var projectDir string
-	st, flags, err := parseCommon("compose sync", ios, args, func(fs *flag.FlagSet) {
-		fs.StringVar(&projectDir, "project-directory", "", "directory to look up hikyo-compose.yaml from (walks up)")
-	})
-	if err != nil {
-		return err
-	}
-	if err := flags.checkNoPositionals("compose sync"); err != nil {
-		return err
-	}
-
-	// (1) Doctor checks run BEFORE the first render; any BLOCKING error finding
-	// refuses without rendering. Findings go to stderr — stdout stays empty. The
-	// server-agreement family (`server_manifest_drift`, `never_rendered`,
-	// `server_stamp_unknown`, `server_unreachable`) is DELIBERATELY EXCLUDED from
-	// the gate: those describe exactly the staleness this sync is about to repair,
-	// so gating on them would brick sync on every publish and every fresh box.
-	// Sync's gate is the LOCAL integrity checks (version floor, format raw, stamp
-	// grammar, token/state modes); the drift stays an error for the doctor VERB,
-	// which reports rather than repairs (finding 11 — this is the interpretation
-	// of the ADR "same checks" sentence offered for human disposition; see the
-	// spellings § Compose delivery).
-	findings, err := composeDoctorGather(ctx, ios, st, flags, projectDir, false)
-	if err != nil {
-		return err
-	}
-	if hasBlockingError(findings) {
-		if rerr := renderComposeFindings(ios.Stderr, FormatTable, findings); rerr != nil {
-			return failf(ExitInternal, "hikyo compose sync: rendering findings: %v", rerr)
-		}
-		return failf(ExitRefused, "hikyo compose sync: doctor found errors; not rendering")
-	}
-
-	// (2) Render (conditional).
-	moved, stack, err := composeRenderCore(ctx, ios, st, flags, projectDir, false)
-	if err != nil {
-		return err
-	}
-
-	// (3) Apply through `docker compose up -d` when a stamp moved, a prior sync
-	// left an apply-pending marker, OR active stamps differ from the durable
-	// last-applied record. The last comparison closes the crash window after the
-	// stamp rename but before a marker write: no active generation can become
-	// permanently unapplied merely because the process died at that boundary.
-	pending := applyPendingExists(stack.stateDir)
-	stamps, err := compose.CurrentStamps(stack.cfgDir)
-	if err != nil {
-		return failf(ExitRefused, "hikyo compose sync: %v", err)
-	}
-	applied, err := loadAppliedStamps(stack.stateDir)
-	if err != nil {
-		return failf(ExitRefused, "hikyo compose sync: %v", err)
-	}
-	if !moved && !pending && !stampsNeedApply(stamps, applied) {
-		return nil
-	}
-	if err := writeApplyPending(stack.stateDir, stamps); err != nil {
-		return failf(ExitInternal, "hikyo compose sync: writing apply-pending marker: %v", err)
-	}
-	if err := dockerComposeUp(ctx, ios, stack.cfgDir); err != nil {
-		return err // marker stays: the next sync retries the apply
-	}
-	if err := writeAppliedStamps(stack.stateDir, stamps); err != nil {
-		return failf(ExitInternal, "hikyo compose sync: recording applied stamps: %v", err)
-	}
-	if err := removeApplyPending(stack.stateDir); err != nil {
-		return failf(ExitInternal, "hikyo compose sync: clearing apply-pending marker: %v", err)
-	}
-	return nil
-}
-
-func dockerComposeUp(ctx context.Context, ios IO, projectDir string) error {
-	bin, found := dockerBinary(ios)
-	if !found {
-		return failf(ExitRefused, "hikyo compose sync: docker not found on PATH; install Docker Compose or set HIKYO_COMPOSE_DOCKER")
-	}
-	cmd := exec.CommandContext(ctx, bin, "compose", "up", "-d")
-	cmd.Dir = projectDir
-	cmd.Stdin = os.Stdin
-	// `compose sync` is a delivery operation: nothing is printed on hikyo stdout
-	// (output ADR). Docker's own stdout is diagnostic, so it is routed to hikyo
-	// STDERR, keeping sync stdout empty (finding 14).
-	cmd.Stdout = ios.Stderr
-	cmd.Stderr = ios.Stderr
-	// The workload credential never reaches a subprocess (finding 1).
-	cmd.Env = sanitizedEnviron()
-	if err := cmd.Run(); err != nil {
-		var ee *exec.ExitError
-		if errors.As(err, &ee) {
-			return failf(ExitInternal, "hikyo compose sync: `docker compose up -d` exited %d", ee.ExitCode())
-		}
-		return failf(ExitInternal, "hikyo compose sync: `docker compose up -d`: %v", err)
-	}
-	return nil
-}
-
-const (
-	applyPendingFile  = "apply-pending"
-	appliedStampsFile = "applied-stamps.json"
-)
-
-// applyPendingExists reports whether a prior sync left an unfinished apply.
-func applyPendingExists(stateDir string) bool {
-	_, err := os.Stat(filepath.Join(stateDir, applyPendingFile))
-	return err == nil
-}
-
-// writeApplyPending records the stamps to apply, atomically, BEFORE docker runs.
-func writeApplyPending(stateDir string, stamps map[string]string) error {
-	data, err := json.Marshal(stamps)
-	if err != nil {
-		return err
-	}
-	return writeFileAtomic0600(filepath.Join(stateDir, applyPendingFile), data)
-}
-
-// loadAppliedStamps returns the last stamp selection successfully handed to
-// Docker. A missing file is an unapplied state, including upgrades from clients
-// predating this record. Malformed state fails loudly rather than skipping an
-// apply on guessed data.
-func loadAppliedStamps(stateDir string) (map[string]string, error) {
-	b, err := os.ReadFile(filepath.Join(stateDir, appliedStampsFile))
-	if err != nil {
-		if errors.Is(err, os.ErrNotExist) {
-			return nil, nil
-		}
-		return nil, fmt.Errorf("reading applied stamps: %w", err)
-	}
-	var stamps map[string]string
-	if err := json.Unmarshal(b, &stamps); err != nil {
-		return nil, fmt.Errorf("parsing applied stamps: %w", err)
-	}
-	if stamps == nil {
-		return nil, errors.New("parsing applied stamps: expected a target-to-stamp object")
-	}
-	for target, stamp := range stamps {
-		if strings.TrimSpace(target) == "" {
-			return nil, errors.New("parsing applied stamps: target name is empty")
-		}
-		if err := crypto.ParseStamp(stamp); err != nil {
-			return nil, fmt.Errorf("parsing applied stamps for %q: %w", target, err)
-		}
-	}
-	return stamps, nil
-}
-
-func writeAppliedStamps(stateDir string, stamps map[string]string) error {
-	for target, stamp := range stamps {
-		if strings.TrimSpace(target) == "" {
-			return errors.New("writing applied stamps: target name is empty")
-		}
-		if err := crypto.ParseStamp(stamp); err != nil {
-			return fmt.Errorf("writing applied stamps for %q: %w", target, err)
-		}
-	}
-	data, err := json.Marshal(stamps)
-	if err != nil {
-		return err
-	}
-	return writeFileAtomic0600(filepath.Join(stateDir, appliedStampsFile), data)
-}
-
-func stampsNeedApply(current, applied map[string]string) bool {
-	return !maps.Equal(current, applied)
-}
-
-// removeApplyPending clears the marker after a successful docker apply.
-func removeApplyPending(stateDir string) error {
-	err := os.Remove(filepath.Join(stateDir, applyPendingFile))
-	if err != nil && !errors.Is(err, os.ErrNotExist) {
-		return err
-	}
-	return nil
-}
-
-// ---------------------------------------------------------------------------
-// hikyo compose doctor
-// ---------------------------------------------------------------------------
-
-func runComposeDoctor(ctx context.Context, ios IO, args []string) error {
-	var (
-		format     string
-		projectDir string
-	)
-	st, flags, err := parseCommon("compose doctor", ios, args, func(fs *flag.FlagSet) {
-		fs.StringVar(&format, "o", "table", "output format: table or json")
-		fs.StringVar(&projectDir, "project-directory", "", "directory to look up hikyo-compose.yaml from (walks up)")
-	})
-	if err != nil {
-		return err
-	}
-	if err := flags.checkNoPositionals("compose doctor"); err != nil {
-		return err
-	}
-	f, err := ParseFormat(format)
-	if err != nil {
-		return err
-	}
-	findings, err := composeDoctorGather(ctx, ios, st, flags, projectDir, true)
-	if err != nil {
-		return err
-	}
-	// Propagate a rendering failure as ExitInternal (finding 15): doctor must not
-	// return success or its findings-derived refusal when its required output
-	// never reached the caller.
-	if rerr := renderComposeFindings(ios.Stdout, f, findings); rerr != nil {
-		return failf(ExitInternal, "hikyo compose doctor: rendering findings: %v", rerr)
-	}
-	if hasErrorFinding(findings) {
-		return failf(ExitRefused, "hikyo compose doctor found errors")
-	}
-	return nil
-}
-
-// composeDoctorGather assembles every input compose.Doctor needs — docker
-// version/config, the raw compose file, managed stamps, generation state, file
-// modes, and server agreement via a conditional fetch — and returns the merged
-// finding list.
-func composeDoctorGather(ctx context.Context, ios IO, st *State, flags commonFlags, projectDir string, includeServerAgreement bool) ([]compose.Finding, error) {
-	stack, err := openComposeStack(st, ios, flags, composeStackOptions{projectDir: projectDir, requireConfig: true})
-	if err != nil {
-		return nil, err
-	}
-
-	// Flush-before-fetch (ops-spec § 6): reconcile pending offline records BEFORE
-	// any doctor network request (the catalogue and agreement fetches), so a POST
-	// always precedes every GET (finding 9). A flush failure is a hard error.
-	if err := stack.flushOffline(ctx); err != nil {
-		return nil, err
-	}
-
-	var findings []compose.Finding
-
-	// Docker version + resolved config.
-	dockerFindings, version, resolvedConfig := doctorDocker(ctx, ios, stack.cfgDir)
-	findings = append(findings, dockerFindings...)
-
-	managed, err := compose.CurrentStamps(stack.cfgDir)
-	if err != nil {
-		return nil, failf(ExitRefused, "compose doctor: %v", err)
-	}
-
-	existingKeyIDs, catFinding := doctorExistingKeyIDs(ctx, stack.client, stack.org, stack.project, stack.cfg)
-	if catFinding != nil {
-		findings = append(findings, *catFinding)
-	}
-	stateEntries, scanFinding := doctorStateEntries(stack.stateDir)
-	if scanFinding != nil {
-		findings = append(findings, *scanFinding)
-	}
-
-	in := compose.DoctorInput{
-		ComposeVersion: version,
-		Config:         resolvedConfig,
-		RawComposeYAML: doctorRawCompose(ios, stack.cfgDir),
-		ManagedStamps:  managed,
-		ConfigTargets:  stack.cfg.Targets,
-		ExistingKeyIDs: existingKeyIDs,
-		StateEntries:   stateEntries,
-		TokenFile:      doctorTokenFile(flags.TokenFile),
-
-		SystemdInvocation:       ios.Env.Getenv("INVOCATION_ID") != "",
-		TokenFromCredentialsDir: tokenFromCredentialsDir(ios, flags.TokenFile),
-	}
-	// runtime_dir must resolve; when it cannot (not root, no XDG_RUNTIME_DIR, no
-	// explicit config), surface it as its own error and do not let the derived
-	// runtime checks fire on an empty path (finding 12).
-	runtimeResolved := stack.runtimeErr == nil
-	if runtimeResolved {
-		in.RuntimeDir = stack.runtimeDir
-		in.RuntimeTmpfs = doctorRuntimeTmpfs(stack.runtimeDir)
-	}
-
-	// Server agreement: feed compose.Doctor the per-target server stamps it needs
-	// (finding 4). The structural check then compares them against the managed
-	// stamp and the label. Only the DOCTOR verb reaches the server; sync repairs
-	// freshness and must not gate on it (finding 11).
-	var serverStamps map[string]string
-	var serverFindings []compose.Finding
-	haveServerStamps := false
-	if includeServerAgreement {
-		serverStamps, serverFindings, haveServerStamps = stack.doctorServerStamps(ctx, managed)
-		in.ServerStamps = serverStamps
-	}
-
-	findings = append(findings, compose.Doctor(in)...)
-	findings = append(findings, serverFindings...)
-
-	// A docker_missing finding makes the version check redundant; drop the
-	// compose.Doctor floor finding so the two do not both fire for one cause.
-	if hasCode(findings, "docker_missing") {
-		findings = dropCode(findings, "compose_version_below_floor")
-	}
-	if !runtimeResolved {
-		findings = dropCode(findings, "runtime_not_tmpfs")
-		findings = dropCode(findings, "runtime_dir_not_absolute")
-		findings = append(findings, compose.Finding{Severity: compose.SeverityError, Code: "runtime_dir_unresolved",
-			Message: fmt.Sprintf("could not resolve a runtime dir: %v", stack.runtimeErr)})
-	}
-	// server_stamp_unknown is compose.Doctor's honest "no server stamp to compare"
-	// finding. Drop it whenever we deliberately did not (or could not) obtain the
-	// server stamps — sync never fetches, and doctor's never_rendered /
-	// server_unreachable / drift cases explain the gap directly.
-	if !haveServerStamps {
-		findings = dropCode(findings, "server_stamp_unknown")
-	}
-
-	sortFindings(findings)
-	return findings, nil
-}
-
-// doctorDocker runs `docker compose version --short` and `docker compose config
-// --format json`. A missing docker binary is docker_missing; a config invocation
-// or JSON parse failure is docker_config_failed (fail closed — a nil config used
-// to silently disable the service checks, finding 12).
-func doctorDocker(ctx context.Context, ios IO, cfgDir string) ([]compose.Finding, string, *compose.ComposeConfig) {
-	bin, found := dockerBinary(ios)
-	if !found {
-		return []compose.Finding{{Severity: compose.SeverityError, Code: "docker_missing",
-			Message: "docker not found on PATH and HIKYO_COMPOSE_DOCKER is unset; path 2 needs Docker Compose ≥ 2.30"}}, "", nil
-	}
-	version, err := runCapture(ctx, ios, bin, cfgDir, "compose", "version", "--short")
-	if err != nil {
-		return []compose.Finding{{Severity: compose.SeverityError, Code: "docker_missing",
-			Message: fmt.Sprintf("`docker compose version` failed: %v", err)}}, "", nil
-	}
-	raw, cerr := runCapture(ctx, ios, bin, cfgDir, "compose", "config", "--format", "json")
-	if cerr != nil {
-		return []compose.Finding{{Severity: compose.SeverityError, Code: "docker_config_failed",
-			Message: fmt.Sprintf("`docker compose config` failed: %v", cerr)}}, strings.TrimSpace(version), nil
-	}
-	parsed, perr := compose.ParseComposeConfig([]byte(raw))
-	if perr != nil {
-		return []compose.Finding{{Severity: compose.SeverityError, Code: "docker_config_failed",
-			Message: fmt.Sprintf("could not parse `docker compose config` JSON: %v", perr)}}, strings.TrimSpace(version), nil
-	}
-	return nil, strings.TrimSpace(version), parsed
-}
-
-// doctorServerStamps performs the conditional agreement fetch and returns the
-// per-target server stamps to feed compose.Doctor's structural check, plus any
-// standalone findings. The bool reports whether server stamps were obtained (so
-// the caller can drop the honest server_stamp_unknown when they were not).
-func (s *composeStack) doctorServerStamps(ctx context.Context, managed map[string]string) (map[string]string, []compose.Finding, bool) {
-	present := s.eligibleCursor(managed)
-	if present == "" {
-		// No eligible cursor: never rendered, or the local render is gone. A full
-		// fetch would be a disclosure, so doctor does not do one.
-		return nil, []compose.Finding{{Severity: compose.SeverityError, Code: "never_rendered",
-			Message: "no eligible cursor: this box has not rendered, or its render is gone; run `hikyo compose render`"}}, false
-	}
-	resp, err := s.fetchDelivery(ctx, renderAcknowledged(s.cfg), present)
-	if err != nil {
-		sev := compose.SeverityError
-		msg := fmt.Sprintf("the server refused the agreement check: %v", err)
-		if isUnavailable(err) {
-			sev = compose.SeverityWarn
-			msg = fmt.Sprintf("could not reach the server to confirm agreement: %v", err)
-		}
-		return nil, []compose.Finding{{Severity: sev, Code: "server_unreachable", Message: msg}}, false
-	}
-	if resp.Current {
-		// The server agrees the presented cursor is current: our managed stamps ARE
-		// the server's, no plaintext crossed the wire.
-		return managed, nil, true
-	}
-	// Drift: the cursor is not current. We cannot always recompute the server's
-	// per-target stamp (a read-only credential holds no secret plaintext), so
-	// report the drift directly and leave the stamps unknown.
-	return nil, []compose.Finding{{Severity: compose.SeverityError, Code: "server_manifest_drift",
-		Message: "the server's current manifest is not the one this box rendered — run `hikyo compose render`"}}, false
 }
 
 // ---------------------------------------------------------------------------
@@ -1661,23 +653,6 @@ func composeRuntimeDir(ios IO, cfg *compose.Config, slug string) (string, bool, 
 		"no runtime directory: not root and XDG_RUNTIME_DIR is unset. Set `runtime_dir` in %s, or run under a session with XDG_RUNTIME_DIR", composeConfigName)
 }
 
-// doctorRuntimeTmpfs reports whether the runtime dir (or its nearest existing
-// ancestor, since the dir may not exist yet) is tmpfs-backed. On non-Linux
-// IsTmpfs returns true, so this never falsely flags there.
-func doctorRuntimeTmpfs(dir string) bool {
-	for dir != "" {
-		if ok, err := compose.IsTmpfs(dir); err == nil {
-			return ok
-		}
-		parent := filepath.Dir(dir)
-		if parent == dir {
-			break
-		}
-		dir = parent
-	}
-	return true
-}
-
 func loadLocalKeys(stateDir string) (*crypto.LocalKeys, error) {
 	keys, err := crypto.LoadOrCreateLocalKey(stateDir)
 	if err != nil {
@@ -1687,353 +662,12 @@ func loadLocalKeys(stateDir string) (*crypto.LocalKeys, error) {
 }
 
 // ---------------------------------------------------------------------------
-// doctor input gathering
-// ---------------------------------------------------------------------------
-
-// doctorRawCompose reads the raw compose file so ${HIKYO_GEN_*:?} is visible
-// (the resolved config discards the required form). COMPOSE_FILE wins when set;
-// otherwise the first conventional name found in the project dir.
-func doctorRawCompose(ios IO, cfgDir string) string {
-	if cf := ios.Env.Getenv("COMPOSE_FILE"); cf != "" {
-		path := cf
-		if !filepath.IsAbs(path) {
-			path = filepath.Join(cfgDir, path)
-		}
-		if b, err := os.ReadFile(path); err == nil {
-			return string(b)
-		}
-		return ""
-	}
-	for _, name := range []string{"compose.yaml", "compose.yml", "docker-compose.yaml", "docker-compose.yml"} {
-		if b, err := os.ReadFile(filepath.Join(cfgDir, name)); err == nil {
-			return string(b)
-		}
-	}
-	return ""
-}
-
-// doctorExistingKeyIDs reads the project key catalogue for the target_key_missing
-// check. A workload credential deliberately CANNOT enumerate the catalogue (it
-// reads values through delivery, not the catalogue), so a UNIFORM not-found /
-// unauthorized there (unauthorized ≡ nonexistent) is not a drift signal — it
-// means the check is not answerable from this credential, and the configured ids
-// are treated as existing (a no-op). Any OTHER failure — transport, 5xx, decode
-// — is NOT swallowed: it fails closed as `catalogue_unavailable` (finding 12),
-// while still treating the ids as existing so target_key_missing does not fire
-// on an unknown set.
-func doctorExistingKeyIDs(ctx context.Context, client *Client, org, project string, cfg *compose.Config) (map[string]bool, *compose.Finding) {
-	allExist := func() map[string]bool {
-		out := map[string]bool{}
-		for _, id := range allTargetKeyIDs(cfg) {
-			out[id] = true
-		}
-		return out
-	}
-	var list apigen.KeyList
-	path := api.PathPrefix + "/orgs/" + url.PathEscape(org) + "/projects/" + url.PathEscape(project) + "/keys"
-	if err := client.Do(ctx, http.MethodGet, path, nil, &list); err != nil {
-		if isNotFound(err) {
-			// Uniform 404/unauthorized: unanswerable from this credential, no-op.
-			return allExist(), nil
-		}
-		return allExist(), &compose.Finding{Severity: compose.SeverityError, Code: "catalogue_unavailable",
-			Message: fmt.Sprintf("could not read the key catalogue to confirm target key ids: %v", err)}
-	}
-	out := make(map[string]bool, len(list.Items))
-	for _, k := range list.Items {
-		out[k.Id] = true
-	}
-	return out, nil
-}
-
-// doctorStateEntries walks the client state dir for mode/ownership checks. A
-// walk or stat error is NOT swallowed: it fails closed as `state_scan_failed`
-// (finding 12) rather than silently reporting an incomplete tree.
-func doctorStateEntries(stateDir string) ([]compose.StateEntry, *compose.Finding) {
-	var entries []compose.StateEntry
-	var scanErr error
-	_ = filepath.WalkDir(stateDir, func(path string, d os.DirEntry, err error) error {
-		if err != nil {
-			// A missing state dir is legitimate (nothing rendered yet); any other
-			// walk error is a real scan failure.
-			if errors.Is(err, os.ErrNotExist) && path == stateDir {
-				return nil
-			}
-			scanErr = fmt.Errorf("%s: %w", path, err)
-			return nil
-		}
-		info, ierr := d.Info()
-		if ierr != nil {
-			scanErr = fmt.Errorf("%s: %w", path, ierr)
-			return nil
-		}
-		entries = append(entries, compose.StateEntry{
-			Path: path, Perm: info.Mode(), IsDir: d.IsDir(), OwnedByEUID: ownedByEUID(info),
-		})
-		return nil
-	})
-	if scanErr != nil {
-		return entries, &compose.Finding{Severity: compose.SeverityError, Code: "state_scan_failed",
-			Message: fmt.Sprintf("could not fully scan the client state dir: %v", scanErr)}
-	}
-	return entries, nil
-}
-
-func doctorTokenFile(tokenFile string) *compose.FileMode {
-	if tokenFile == "" {
-		return nil
-	}
-	info, err := os.Stat(tokenFile)
-	if err != nil {
-		return nil
-	}
-	return &compose.FileMode{Perm: info.Mode(), OwnedByEUID: ownedByEUID(info)}
-}
-
-func tokenFromCredentialsDir(ios IO, tokenFile string) bool {
-	dir := ios.Env.Getenv("CREDENTIALS_DIRECTORY")
-	if dir == "" || tokenFile == "" {
-		return false
-	}
-	abs, err := filepath.Abs(tokenFile)
-	if err != nil {
-		return false
-	}
-	rel, err := filepath.Rel(dir, abs)
-	return err == nil && !strings.HasPrefix(rel, "..")
-}
-
-// ---------------------------------------------------------------------------
-// doctor rendering
-// ---------------------------------------------------------------------------
-
-type composeFinding struct {
-	Status  string `json:"status"`
-	Check   string `json:"check"`
-	Message string `json:"message"`
-}
-
-type composeDoctorReport struct {
-	Status   string           `json:"status"`
-	Findings []composeFinding `json:"findings"`
-}
-
-func renderComposeFindings(w io.Writer, f Format, findings []compose.Finding) error {
-	report := composeDoctorReport{Status: "ok", Findings: []composeFinding{}}
-	rows := make([][]string, 0, len(findings))
-	for _, fd := range findings {
-		report.Findings = append(report.Findings, composeFinding{Status: string(fd.Severity), Check: fd.Code, Message: fd.Message})
-		rows = append(rows, []string{string(fd.Severity), fd.Code, fd.Message})
-		switch fd.Severity {
-		case compose.SeverityError:
-			report.Status = "error"
-		case compose.SeverityWarn:
-			if report.Status == "ok" {
-				report.Status = "warning"
-			}
-		}
-	}
-	if len(rows) == 0 {
-		rows = append(rows, []string{"ok", "compose", "no findings"})
-	}
-	return Render(w, f, Table{Columns: []string{"STATUS", "CHECK", "MESSAGE"}, Rows: rows, JSON: report})
-}
-
-// ---------------------------------------------------------------------------
-// docker seam + capture
-// ---------------------------------------------------------------------------
-
-// dockerBinary resolves the docker executable. HIKYO_COMPOSE_DOCKER overrides
-// PATH resolution — the test seam, documented in the handoff/package doc only,
-// deliberately kept out of the help text.
-func dockerBinary(ios IO) (string, bool) {
-	if v := ios.Env.Getenv("HIKYO_COMPOSE_DOCKER"); v != "" {
-		return v, true
-	}
-	bin, err := exec.LookPath("docker")
-	if err != nil {
-		return "", false
-	}
-	return bin, true
-}
-
-func runCapture(ctx context.Context, ios IO, bin, dir string, args ...string) (string, error) {
-	_ = ios
-	cmd := exec.CommandContext(ctx, bin, args...)
-	cmd.Dir = dir
-	// The workload credential never reaches a subprocess (finding 1).
-	cmd.Env = sanitizedEnviron()
-	var out bytes.Buffer
-	cmd.Stdout = &out
-	cmd.Stderr = io.Discard
-	if err := cmd.Run(); err != nil {
-		return "", err
-	}
-	return out.String(), nil
-}
-
-// ---------------------------------------------------------------------------
 // small pure helpers
 // ---------------------------------------------------------------------------
-
-func splitCSV(s string) []string {
-	if strings.TrimSpace(s) == "" {
-		return nil
-	}
-	parts := strings.Split(s, ",")
-	out := make([]string, 0, len(parts))
-	for _, p := range parts {
-		if t := strings.TrimSpace(p); t != "" {
-			out = append(out, t)
-		}
-	}
-	return out
-}
 
 func isUnavailable(err error) bool {
 	var ce *Error
 	return asCLIError(err, &ce) && ce.Code == ExitUnavailable
-}
-
-func unrevealedSecrets(keys []apigen.DeliveredKey) []string {
-	var missing []string
-	for _, k := range keys {
-		if isUnrevealedSecret(k) {
-			missing = append(missing, k.Name)
-		}
-	}
-	sort.Strings(missing)
-	return missing
-}
-
-func isUnrevealedSecret(k apigen.DeliveredKey) bool {
-	return k.Presence == apigen.DeliveredKeyPresenceSet &&
-		k.Classification == apigen.KeyClassificationSecret && k.Value == nil
-}
-
-func liveRenderInput(cfg *compose.Config, configOnly bool, keys []apigen.DeliveredKey) compose.RenderInput {
-	rows := make([]compose.RenderSourceRow, 0, len(keys))
-	for _, key := range keys {
-		row := compose.RenderSourceRow{
-			KeyID: key.KeyId, Name: key.Name, Classification: string(key.Classification),
-		}
-		switch {
-		case !configOnly && isUnrevealedSecret(key):
-			row.State = compose.RenderRowUnrevealedSecret
-		case key.Value == nil:
-			row.State = compose.RenderRowNoValue
-		default:
-			row.State = compose.RenderRowValued
-			row.Value = *key.Value
-		}
-		rows = append(rows, row)
-	}
-	return renderInput(cfg, configOnly, compose.AbsentKeyRefuseNotDelivered, rows)
-}
-
-func offlineRenderInput(cfg *compose.Config, configOnly bool, rows []compose.SnapshotRow) compose.RenderInput {
-	sourceRows := make([]compose.RenderSourceRow, 0, len(rows))
-	for _, row := range rows {
-		sourceRows = append(sourceRows, compose.RenderSourceRow{
-			KeyID: row.KeyID, Name: row.Name, Classification: row.Classification,
-			State: compose.RenderRowValued, Value: row.Value,
-		})
-	}
-	return renderInput(cfg, configOnly, compose.AbsentKeyRefuseNotInSnapshot, sourceRows)
-}
-
-func renderInput(cfg *compose.Config, configOnly bool, fullProjectionPolicy compose.AbsentKeyPolicy, rows []compose.RenderSourceRow) compose.RenderInput {
-	absentKeys := fullProjectionPolicy
-	if configOnly {
-		absentKeys = compose.AbsentKeySkip
-	}
-	targets := make([]compose.RenderTarget, 0, len(cfg.Targets))
-	for _, name := range cfg.TargetNames() {
-		target := cfg.Targets[name]
-		targets = append(targets, compose.RenderTarget{
-			Name: name, KeyIDs: append([]string(nil), target.Keys...),
-			AcknowledgeLoaderControl: append([]string(nil), target.AcknowledgeLoaderControl...),
-		})
-	}
-	return compose.RenderInput{AbsentKeys: absentKeys, Targets: targets, Rows: rows}
-}
-
-func renderRefusalMessages(refusals []compose.RenderRefusal) []string {
-	messages := make([]string, 0, len(refusals))
-	for _, refusal := range refusals {
-		var detail string
-		switch refusal.Kind {
-		case compose.RenderRefusalKeyNotDelivered:
-			detail = "key id was not delivered by the server"
-		case compose.RenderRefusalKeyNotInSnapshot:
-			detail = "not present in the last snapshot"
-		case compose.RenderRefusalSecretUnrevealed:
-			detail = "secret has no value — " + machineRevealOptIn
-		case compose.RenderRefusalLoaderControl:
-			detail = "loader-control key not acknowledged (add it to this target's acknowledge_loader_control)"
-		case compose.RenderRefusalEncoding:
-			detail = refusal.Reason
-		default:
-			detail = fmt.Sprintf("unknown render refusal %q", refusal.Kind)
-		}
-		messages = append(messages, fmt.Sprintf("%s: %s: %s", refusal.Target, refusal.Key, detail))
-	}
-	sort.Strings(messages)
-	return messages
-}
-
-func deliveredValues(keys []apigen.DeliveredKey) map[string]string {
-	out := map[string]string{}
-	for _, k := range keys {
-		if k.Value != nil {
-			out[k.Name] = *k.Value
-		}
-	}
-	return out
-}
-
-func deliveredRows(keys []apigen.DeliveredKey) []compose.SnapshotRow {
-	var rows []compose.SnapshotRow
-	for _, k := range keys {
-		if k.Value == nil {
-			continue
-		}
-		rows = append(rows, compose.SnapshotRow{Name: k.Name, KeyID: k.KeyId, Classification: string(k.Classification), Value: *k.Value})
-	}
-	return rows
-}
-
-// rowNames returns the snapshot row names (for the loader-control check).
-func rowNames(rows []compose.SnapshotRow) []string {
-	out := make([]string, 0, len(rows))
-	for _, r := range rows {
-		out = append(out, r.Name)
-	}
-	return out
-}
-
-func rowsToValues(rows []compose.SnapshotRow) map[string]string {
-	out := make(map[string]string, len(rows))
-	for _, r := range rows {
-		out[r.Name] = r.Value
-	}
-	return out
-}
-
-func mapKeys(m map[string]string) []string {
-	out := make([]string, 0, len(m))
-	for k := range m {
-		out = append(out, k)
-	}
-	sort.Strings(out)
-	return out
-}
-
-func canonicalRows(rows []compose.SnapshotRow) []byte {
-	sorted := append([]compose.SnapshotRow(nil), rows...)
-	sort.Slice(sorted, func(i, j int) bool { return sorted[i].Name < sorted[j].Name })
-	data, _ := json.Marshal(sorted)
-	return data
 }
 
 func allTargetKeyIDs(cfg *compose.Config) []string {
@@ -2100,194 +734,4 @@ func deliveryProjection(keys []apigen.DeliveredKey) []string {
 		}
 	}
 	return proj
-}
-
-func hasErrorFinding(findings []compose.Finding) bool {
-	return hasSeverity(findings, compose.SeverityError)
-}
-
-// syncRepairableCodes is the family sync's pre-render gate excludes: these are
-// the findings sync's own render REPAIRS, not the local integrity it must not
-// proceed past (finding 11). Two groups: the server-agreement freshness family
-// (the staleness sync exists to reconcile), and the runtime-generation family
-// (`generation_absent`/`generation_incomplete` — a wiped or torn tmpfs copy the
-// render re-materialises, R1-10). Gating on either would brick sync on exactly
-// the state it is invoked to fix (every publish, every fresh box, every reboot
-// that lost the tmpfs).
-var syncRepairableCodes = map[string]bool{
-	"server_manifest_drift": true,
-	"never_rendered":        true,
-	"server_stamp_unknown":  true,
-	"server_unreachable":    true,
-	"generation_absent":     true,
-	"generation_incomplete": true,
-}
-
-// hasBlockingError reports whether any error finding OUTSIDE the sync-repairable
-// family is present — the set sync gates on before rendering.
-func hasBlockingError(findings []compose.Finding) bool {
-	for _, f := range findings {
-		if f.Severity == compose.SeverityError && !syncRepairableCodes[f.Code] {
-			return true
-		}
-	}
-	return false
-}
-
-// isNotFound reports the uniform not-found/unauthorized response (unauthorized ≡
-// nonexistent), the only catalogue answer doctor treats as "unanswerable".
-func isNotFound(err error) bool {
-	var ce *Error
-	return asCLIError(err, &ce) && (ce.Code == ExitNotFound || ce.Code == ExitAuth)
-}
-
-// sanitizedEnviron returns the process environment with the workload credential
-// (HIKYO_TOKEN) removed: it is the ONLY credential-transport env var (the token
-// file is a flag, not an env var), and it must never reach the child or any
-// subprocess the CLI spawns (finding 1). Building the child env from this — not
-// os.Environ() — means the credential was never present, not stripped after.
-//
-// CREDENTIALS_DIRECTORY is deliberately NOT stripped (R1-1, accepted): it is a
-// directory PATH, not a secret, and access to the files under it is per-open and
-// uid-gated by systemd — so stripping the variable protects nothing, while
-// stripping it would break a child that legitimately consumes its own systemd
-// credentials.
-func sanitizedEnviron() []string {
-	src := os.Environ()
-	out := make([]string, 0, len(src))
-	for _, e := range src {
-		if strings.HasPrefix(e, "HIKYO_TOKEN=") {
-			continue
-		}
-		out = append(out, e)
-	}
-	return out
-}
-
-// resolveChildCommand resolves the command after `--`, distinguishing 127 (not
-// found) from 126 (found but not executable). exec.LookPath alone is
-// insufficient: for a BARE name it swallows a found-but-non-executable file as
-// ErrNotFound (verified on darwin), so that case is detected by scanning PATH; a
-// command containing a path separator is stat'd directly and LookPath surfaces
-// fs.ErrPermission (finding 13).
-//
-// A command resolved ONLY via a relative PATH entry (`.` or any non-absolute
-// dir) returns exec.ErrDot: Go deliberately refuses to run cwd-controlled code
-// implicitly, and so do we — treat it as NOT FOUND (127), naming the relative
-// entry, never execute it (NEW-1).
-func resolveChildCommand(command string) (string, error) {
-	path, err := exec.LookPath(command)
-	if err == nil {
-		return path, nil
-	}
-	if errors.Is(err, exec.ErrDot) {
-		return "", failf(ExitCommandNotFound, "hikyo run: %s: resolved only via the relative PATH entry %q; refusing to execute cwd-controlled code — use an absolute path or a PATH of absolute directories", command, filepath.Dir(path))
-	}
-	if errors.Is(err, fs.ErrPermission) {
-		return "", failf(ExitCommandNotExecutable, "hikyo run: %s: found but not executable", command)
-	}
-	if !strings.ContainsRune(command, filepath.Separator) && pathHasNonExecutable(command) {
-		return "", failf(ExitCommandNotExecutable, "hikyo run: %s: found on PATH but not executable", command)
-	}
-	return "", failf(ExitCommandNotFound, "hikyo run: %s: command not found", command)
-}
-
-// pathHasNonExecutable reports whether a bare command name matches a regular
-// file on PATH that lacks execute permission. It reads the PROCESS PATH (the
-// same one exec.LookPath consulted), not an injected env, so the two agree.
-func pathHasNonExecutable(command string) bool {
-	// Windows has no 0o111 execute bit (executability is by extension via
-	// PATHEXT), so this Unix-mode scan would misclassify; LookPath already
-	// classifies correctly there.
-	if runtime.GOOS == "windows" {
-		return false
-	}
-	for _, dir := range filepath.SplitList(os.Getenv("PATH")) {
-		if dir == "" {
-			dir = "."
-		}
-		fi, err := os.Stat(filepath.Join(dir, command))
-		if err != nil || fi.IsDir() {
-			continue
-		}
-		if fi.Mode().Perm()&0o111 == 0 {
-			return true
-		}
-	}
-	return false
-}
-
-func hasSeverity(findings []compose.Finding, sev compose.Severity) bool {
-	for _, f := range findings {
-		if f.Severity == sev {
-			return true
-		}
-	}
-	return false
-}
-
-func hasCode(findings []compose.Finding, code string) bool {
-	for _, f := range findings {
-		if f.Code == code {
-			return true
-		}
-	}
-	return false
-}
-
-func dropCode(findings []compose.Finding, code string) []compose.Finding {
-	out := findings[:0]
-	for _, f := range findings {
-		if f.Code != code {
-			out = append(out, f)
-		}
-	}
-	return out
-}
-
-func sortFindings(findings []compose.Finding) {
-	sort.SliceStable(findings, func(i, j int) bool {
-		if findings[i].Code != findings[j].Code {
-			return findings[i].Code < findings[j].Code
-		}
-		return findings[i].Message < findings[j].Message
-	})
-}
-
-// writeFileAtomic0600 writes data to a temp file in the same dir and renames it
-// into place (0600), creating the directory 0700 if needed.
-func writeFileAtomic0600(path string, data []byte) error {
-	dir := filepath.Dir(path)
-	if err := os.MkdirAll(dir, 0o700); err != nil {
-		return err
-	}
-	tmp, err := os.CreateTemp(dir, ".tmp-"+filepath.Base(path)+"-*")
-	if err != nil {
-		return err
-	}
-	tmpName := tmp.Name()
-	committed := false
-	defer func() {
-		if !committed {
-			tmp.Close()
-			_ = os.Remove(tmpName)
-		}
-	}()
-	if err := tmp.Chmod(0o600); err != nil {
-		return err
-	}
-	if _, err := tmp.Write(data); err != nil {
-		return err
-	}
-	if err := tmp.Sync(); err != nil {
-		return err
-	}
-	if err := tmp.Close(); err != nil {
-		return err
-	}
-	if err := os.Rename(tmpName, path); err != nil {
-		return err
-	}
-	committed = true
-	return nil
 }
