--- plans/README.md
@@ -1,18 +1,0 @@
-# Animation plans
-
-| # | Plan | Severity | Status |
-| --- | --- | --- | --- |
-| 001 | [Reveal flow steps in order](001-ordered-flow-step-reveal.md) | LOW | DONE |
-
-## Recommended execution order
-
-1. Execute plan 001. It is independent and has no prerequisites.
-
-## Dependencies
-
-None.
-
-## Scope
-
-These plans cover landing-page motion only. Documentation navigation and
-documentation UI are excluded.
